#1290 Modularization Alignment
