+++ Source/Prism/Modularity/CyclicDependencyFoundException.Desktop.cs
