+++ Source/Prism/Modularity/CyclicDependencyFoundException.cs
