+++ Source/Prism/Modularity/DuplicateModuleException.Desktop.cs
