+++ Source/Prism/Modularity/DuplicateModuleException.cs
