+++ Source/Prism/Modularity/IModuleCatalogItem.cs
