+++ Source/Prism/Modularity/IModuleInitializer.cs
