+++ Source/Prism/Modularity/IModuleManager.cs
@@ -18,12 +18,14 @@ namespace Prism.Modularity
         /// Loads and initializes the module on the <see cref="ModuleCatalog"/> with the name <paramref name="moduleName"/>.
         /// </summary>
         /// <param name="moduleName">Name of the module requested for initialization.</param>
-        void LoadModule(string moduleName);       
+        void LoadModule(string moduleName);
 
+#if NET45
         /// <summary>
         /// Raised repeatedly to provide progress as modules are downloaded.
         /// </summary>
         event EventHandler<ModuleDownloadProgressChangedEventArgs> ModuleDownloadProgressChanged;
+#endif
 
         /// <summary>
         /// Raised when a module is loaded or fails to load.
