+++ Source/Prism/Modularity/ModularityException.Desktop.cs
