+++ Source/Prism/Modularity/ModularityException.cs
