+++ Source/Prism/Modularity/ModuleInfo.Desktop.cs
