+++ Source/Prism/Modularity/ModuleInitializeException.Desktop.cs
