+++ Source/Prism/Modularity/ModuleInitializeException.cs
