+++ Source/Prism/Modularity/ModuleNotFoundException.Desktop.cs
