+++ Source/Prism/Modularity/ModuleNotFoundException.cs
