+++ Source/Prism/Modularity/ModuleTypeLoaderNotFoundException.cs
