+++ Source/Prism/Modularity/ModuleTypeLoadingException.Desktop.cs
