+++ Source/Prism/Modularity/ModuleTypeLoadingException.cs
