+++ Source/Wpf/Prism.Wpf.Tests/Modularity/ModuleCatalogXaml/SimpleModuleCatalog.xaml
@@ -1,19 +1,20 @@
 ﻿<Modularity:ModuleCatalog xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-               xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                xmlns:sys="clr-namespace:System;assembly=mscorlib"
-               xmlns:Modularity="clr-namespace:Prism.Modularity;assembly=Prism.Wpf">
+                          xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                          xmlns:sys="clr-namespace:System;assembly=mscorlib"
+                          xmlns:coreModularity="clr-namespace:Prism.Modularity;assembly=Prism"
+                          xmlns:Modularity="clr-namespace:Prism.Modularity;assembly=Prism.Wpf">
     <Modularity:ModuleInfoGroup Ref="ModuleGroup1" InitializationMode="OnDemand">
-        <Modularity:ModuleInfo ModuleName="Module1" ModuleType="Module1Type" />
+        <coreModularity:ModuleInfo ModuleName="Module1" ModuleType="Module1Type" />
     </Modularity:ModuleInfoGroup>
     <Modularity:ModuleInfoGroup Ref="ModuleGroup2" InitializationMode="WhenAvailable">
-        <Modularity:ModuleInfo ModuleName="Module2InModuleGroup2" ModuleType="Module2Type">
-            <Modularity:ModuleInfo.DependsOn>
+        <coreModularity:ModuleInfo ModuleName="Module2InModuleGroup2" ModuleType="Module2Type">
+            <coreModularity:ModuleInfo.DependsOn>
                 <sys:String>ModuleW</sys:String>
-            </Modularity:ModuleInfo.DependsOn>
-        </Modularity:ModuleInfo>
-        <Modularity:ModuleInfo ModuleName="Module3InModuleGroup2" ModuleType="ModuleType">
-        </Modularity:ModuleInfo>
+            </coreModularity:ModuleInfo.DependsOn>
+        </coreModularity:ModuleInfo>
+        <coreModularity:ModuleInfo ModuleName="Module3InModuleGroup2" ModuleType="ModuleType">
+        </coreModularity:ModuleInfo>
     </Modularity:ModuleInfoGroup>
-    <Modularity:ModuleInfo Ref="Module3" ModuleName="Module3" ModuleType="Module3Type" />
+    <coreModularity:ModuleInfo Ref="Module3" ModuleName="Module3" ModuleType="Module3Type" />
 </Modularity:ModuleCatalog>
