+++ Source/Wpf/Prism.Wpf/Properties/Resources.resx
@@ -171,24 +171,6 @@
   <data name="FailedToGetType" xml:space="preserve">
     <value>Unable to retrieve the module type {0} from the loaded assemblies.  You may need to specify a more fully-qualified type name.</value>
   </data>
-  <data name="FailedToLoadModule" xml:space="preserve">
-    <value>An exception occurred while initializing module '{0}'. 
-    - The exception message was: {2}
-    - The Assembly that the module was trying to be loaded from was:{1}
-    Check the InnerException property of the exception for more information. If the exception occurred while creating an object in a DI container, you can exception.GetRootException() to help locate the root cause of the problem. 
-  </value>
-  </data>
-  <data name="FailedToLoadModuleNoAssemblyInfo" xml:space="preserve">
-    <value>An exception occurred while initializing module '{0}'. 
-    - The exception message was: {1}
-    Check the InnerException property of the exception for more information. If the exception occurred 
-    while creating an object in a DI container, you can exception.GetRootException() to help locate the 
-    root cause of the problem. </value>
-  </data>
-  <data name="FailedToRetrieveModule" xml:space="preserve">
-    <value>Failed to load type for module {0}. 
-Error was: {1}.</value>
-  </data>
   <data name="HostControlCannotBeNull" xml:space="preserve">
     <value>HostControl cannot have null value when behavior attaches. </value>
   </data>
