--- Source/Xamarin/Prism.Forms/Modularity/IModuleInitializer.cs
@@ -1,14 +1,0 @@
-﻿namespace Prism.Modularity
-{
-    /// <summary>
-    /// Declares a service which initializes the modules into the application.
-    /// </summary>
-    public interface IModuleInitializer
-    {
-        /// <summary>
-        /// Initializes the specified module.
-        /// </summary>
-        /// <param name="moduleInfo">The module to initialize</param>
-        void Initialize(ModuleInfo moduleInfo);
-    }
-}
