--- Source/Xamarin/Prism.Forms/Modularity/IModuleManager.cs
@@ -1,21 +1,0 @@
-﻿using System;
-
-namespace Prism.Modularity
-{
-    /// <summary>
-    /// Defines the interface for the service that will retrieve and initialize the application's modules.
-    /// </summary>
-    public interface IModuleManager
-    {
-        /// <summary>
-        /// Loads and initializes the module on the <see cref="ModuleCatalog"/> with the name <paramref name="moduleName"/>.
-        /// </summary>
-        /// <param name="moduleName">Name of the module requested for initialization.</param>
-        void LoadModule(string moduleName);
-
-        /// <summary>
-        /// Initializes the modules in the <see cref="ModuleCatalog"/>.
-        /// </summary>
-        void Run();
-    }
-}
