+++ Source/Xamarin/Prism.Forms/Modularity/ModuleInitializer.cs
@@ -14,7 +14,7 @@ namespace Prism.Modularity
 
         public void Initialize(ModuleInfo moduleInfo)
         {
-            var module = CreateModule(moduleInfo.ModuleType);
+            var module = CreateModule(Type.GetType(moduleInfo.ModuleType));
             if (module != null)
             {
                 module.RegisterTypes(_container);
